+++ feature/epics/domain/src/main/java/com/grappim/taigamobile/feature/epics/domain/Epic.kt
@@ -13,7 +13,7 @@ data class Epic(
     val version: Long,
     val createdDateTime: LocalDateTime,
     val title: String,
-    val ref: Int,
+    val ref: Long,
     val status: Statuses?,
     val assignee: User? = null,
     val project: Project,
