+++ feature/issues/ui/src/main/java/com/grappim/taigamobile/feature/issues/ui/details/IssueDetailsNavDestination.kt
@@ -4,8 +4,8 @@ import androidx.navigation.NavController
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class IssueDetailsNavDestination(val taskId: Long, val ref: Int)
+data class IssueDetailsNavDestination(val taskId: Long, val ref: Long)
 
-fun NavController.navigateToIssueDetails(taskId: Long, ref: Int) {
+fun NavController.navigateToIssueDetails(taskId: Long, ref: Long) {
     navigate(route = IssueDetailsNavDestination(taskId = taskId, ref = ref))
 }
