+++ feature/tasks/ui/src/main/java/com/grappim/taigamobile/feature/tasks/ui/TaskDetailsNavDestination.kt
@@ -4,8 +4,8 @@ import androidx.navigation.NavController
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class TaskDetailsNavDestination(val taskId: Long, val ref: Int)
+data class TaskDetailsNavDestination(val taskId: Long, val ref: Long)
 
-fun NavController.navigateToTask(taskId: Long, ref: Int) {
+fun NavController.navigateToTask(taskId: Long, ref: Long) {
     navigate(route = TaskDetailsNavDestination(taskId, ref))
 }
