+++ feature/workitem/domain/src/main/java/com/grappim/taigamobile/feature/workitem/domain/EpicShortInfo.kt
@@ -1,3 +1,3 @@
 package com.grappim.taigamobile.feature.workitem.domain
 
-data class EpicShortInfo(val id: Long, val title: String, val ref: Int, val color: String)
+data class EpicShortInfo(val id: Long, val title: String, val ref: Long, val color: String)
